Set correct image during PR checks
The app-interface config defines the image as quay.io/redhat-services-prod, so
the IMAGE must first match that value in order to later override it in the
extra options.

For components, it looks like bonfire uses the IMAGE value from each component repository,
not the IMAGE value in the app-interface deploy.yml.

Since the image forccx-smart-proxy as defined in the git repository is quay.io/cloudservices,
it is trying to find an image in quay.io/cloudservices that matches the commit hash of the
default branch on the repository. Build jobs for that repository were disabled, so there are no
recent images there.

That's the reason for setting the ccx-smart-proxy image to quay.io/redhat-services-prod.