[CCXDEV-14562] New table for runtimes heartbeats (#2080)
* New table for runtimes heartbeats

A migration included in the DVO ones (for simplicity) creates a new table
to store runtimes heartbeats data (basically ID and timestamp).

Included in the DVO ones to avoid creating new migration schema and another entry point. In addition, this table will be populated by DVO-writer (a modified version of the service) and updated by a new service.