Use latest image of go-toolset for building binary
As we are aiming for automatic version bumping and Konflux MRs are pilling up, this change aims to reduce the load with similar results. In addition, as Konflux might change tags with go version, e.g. 1.23 to 9.6, we are not loosing much relevenat information anyways. Finally, we expect that any imcompatibility between the code and the image used to build it is manifested in the CI.