Avoid initializing OCP recommendations table
When there is an issue with the db-writer container not initializing the DBs,
dvo-writer fails with error "no such table: public.migration_info" and
message "unable to check DB migration version"

It seems caused by the code checking both OCP and DVO storages. However,
for DVO writer OCP recommendations storage is not needed.