Move DB definition from aggregator to DVO writer
Currently the DB for aggregator, db-writer and dvo-writer is defined in
clowdapp.yaml which defines the first 2 services. That creates a dependency
that dvo-writer requires the former to spawn.
However, the reality is that dvo-writer can be spawned independently
(after https://github.com/RedHatInsights/insights-results-aggregator/pull/2303 or
https://github.com/RedHatInsights/insights-results-aggregator/pull/2304 are merged).
Thus, in order to change that dependency, we are moving the DB defintion from clowdapp.yaml
to dvo-writer.yaml and use the sharedDbAppName on clowdapp.yaml

We will need to double check that this won't create any issues on stage/prod DB.

One alternative that might be worth considering is splitting clowdapp.yaml
into one file per deployment it defines. Or alternatively move dvo-writer to that clowdapp.